refactor: renomeia classes para pascalcase
