+++ src/core/domain/entities/log.js
@@ -1,6 +1,6 @@
 import Sequelize from "sequelize";
 
-import db from '../../../config/database.js'
+import db from '../../../config/DataBaseConfig.js'
 
 const LogEntity = db.sequelize.define('LogEntity',
     {
